refactor handlers to commands
