+++ src/commands/handleAutoPost.js
@@ -1,4 +1,9 @@
-export const handleAutoPost = async (interaction, action, client, redisClient) => {
+export const handleAutoPost = async (
+  interaction,
+  action,
+  client,
+  redisClient,
+) => {
   const channel = client.channels.cache.get(interaction.channelId);
   const surveyName = interaction.options.getString("survey");
   const key = "auto-post-surveys";
