+++ src/commands/handleInfo.js
