+++ src/commands/handleView.js
